requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.7
django-appconf==1.1.0
django-select2==8.3.0
pillow==11.1.0
python-barcode==0.15.1
reportlab==4.3.1
sqlparse==0.5.3
tzdata==2025.1
